refactor(service): CommandeClientServiceImpl

+++ src/main/java/com/tid/StockMaster/services/impl/CommandeClientServiceImpl.java
@@ -131,6 +131,10 @@ public class CommandeClientServiceImpl implements CommandeClientService {
             log.error("Commande client ID is NULL");
             return;
         }
+        List<LigneCommandeClient> ligneCommandeClients = ligneCommandeClientRepository.findAllByCommandeClientId(id);
+        if(!ligneCommandeClients.isEmpty()){
+            throw new InvalidOperationException("Impossible de supprimmer la commande client",ErrorCodes.COMMANDE_CLIENT_ALREADY_IN_USE);
+        }
         commandeClientRepository.deleteById(id);
     }
